backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "pyshacl>=0.30.1",
    "python-dotenv>=1.1.1",
    "rdflib>=7.2.1",
    "uvicorn>=0.37.0",
]
